+++ src/Casting/Sexadecimal/Round.php
@@ -4,8 +4,6 @@ namespace MarcoConsiglio\Goniometry\Casting\Sexadecimal;
 use MarcoConsiglio\Goniometry\Casting\Sexagesimal;
 use MarcoConsiglio\Goniometry\Interfaces\Casting\ToSexadecimal;
 use MarcoConsiglio\Goniometry\Interfaces\SexadecimalValue;
-use MarcoConsiglio\Goniometry\SexadecimalDegrees;
-use RoundingMode;
 
 /**
  * Round a sexadecimal value.