+++ tests/Feature/AngularDistanceTest.php
@@ -2,21 +2,33 @@
 namespace MarcoConsiglio\Goniometry\Tests\Feature;
 
 use MarcoConsiglio\FakerPhpNumberHelpers\NextFloat;
+use MarcoConsiglio\Goniometry\Angle;
 use MarcoConsiglio\Goniometry\AngularDistance;
+use MarcoConsiglio\Goniometry\Builders\Angle\FromSexagesimal as AngleFromSexagesimal;
 use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromSexadecimal;
+use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromSexagesimal;
 use MarcoConsiglio\Goniometry\Casting\Sexadecimal\Round;
 use MarcoConsiglio\Goniometry\Casting\Sexagesimal;
 use MarcoConsiglio\Goniometry\Degrees;
 use MarcoConsiglio\Goniometry\Minutes;
+use MarcoConsiglio\Goniometry\Random\Generator\Degrees as DegreesGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\FloatGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\Minutes as MinutesGenerator;
 use MarcoConsiglio\Goniometry\Random\Generator\NegativeSexadecimal as NegativeSexadecimalGenerator;
 use MarcoConsiglio\Goniometry\Random\Generator\PositiveSexadecimal as PositiveSexadecimalGenerator;
 use MarcoConsiglio\Goniometry\Random\Generator\RelativeSexadecimal as RelativeSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\Seconds as SecondsGenerator;
+use MarcoConsiglio\Goniometry\Random\SecondsRange;
+use MarcoConsiglio\Goniometry\Random\Validator\Degrees as DegreesValidator;
 use MarcoConsiglio\Goniometry\Random\Validator\FloatValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\Minutes as MinutesValidator;
 use MarcoConsiglio\Goniometry\Random\Validator\NegativeSexadecimal as NegativeSexadecimalValidator;
 use MarcoConsiglio\Goniometry\Random\Validator\PositiveSexadecimal as PositiveSexadecimalValidator;
 use MarcoConsiglio\Goniometry\Random\Validator\RelativeSexadecimal as RelativeSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\Seconds as SecondsValidator;
 use MarcoConsiglio\Goniometry\Seconds;
 use MarcoConsiglio\Goniometry\SexadecimalAngularDistance;
+use MarcoConsiglio\Goniometry\SexadecimalDegrees;
 use MarcoConsiglio\Goniometry\SexagesimalDegrees;
 use MarcoConsiglio\Goniometry\Tests\TestCase;
 use MarcoConsiglio\Goniometry\Traits\WithAngleFaker;
@@ -26,24 +38,50 @@ use PHPUnit\Framework\Attributes\UsesClass;
 use PHPUnit\Framework\Attributes\UsesTrait;
 
 #[CoversClass(AngularDistance::class)]
-#[UsesClass(FromSexadecimal::class)]
-#[UsesClass(Round::class)]
-#[UsesClass(Sexagesimal::class)]
+#[UsesClass(Angle::class)]
+#[UsesClass(AngleFromSexagesimal::class)]
 #[UsesClass(Degrees::class)]
-#[UsesClass(Minutes::class)]
-#[UsesClass(PositiveSexadecimalGenerator::class)]
-#[UsesClass(RelativeSexadecimalGenerator::class)]
+#[UsesClass(DegreesGenerator::class)]
+#[UsesClass(DegreesValidator::class)]
+#[UsesClass(FloatGenerator::class)]
 #[UsesClass(FloatValidator::class)]
-#[UsesClass(PositiveSexadecimalValidator::class)]
-#[UsesClass(RelativeSexadecimalValidator::class)]
-#[UsesClass(Seconds::class)]
-#[UsesClass(SexadecimalAngularDistance::class)]
-#[UsesClass(SexagesimalDegrees::class)]
+#[UsesClass(FromSexadecimal::class)]
+#[UsesClass(FromSexagesimal::class)]
+#[UsesClass(Minutes::class)]
+#[UsesClass(MinutesGenerator::class)]
+#[UsesClass(MinutesValidator::class)]
 #[UsesClass(NegativeSexadecimalGenerator::class)]
 #[UsesClass(NegativeSexadecimalValidator::class)]
+#[UsesClass(PositiveSexadecimalGenerator::class)]
+#[UsesClass(PositiveSexadecimalValidator::class)]
+#[UsesClass(RelativeSexadecimalGenerator::class)]
+#[UsesClass(RelativeSexadecimalValidator::class)]
+#[UsesClass(Round::class)]
+#[UsesClass(Seconds::class)]
+#[UsesClass(SecondsGenerator::class)]
+#[UsesClass(SecondsRange::class)]
+#[UsesClass(SecondsValidator::class)]
+#[UsesClass(SexadecimalAngularDistance::class)]
+#[UsesClass(SexadecimalDegrees::class)]
+#[UsesClass(Sexagesimal::class)]
+#[UsesClass(SexagesimalDegrees::class)]
 #[UsesTrait(WithAngleFaker::class)]
 class AngularDistanceTest extends TestCase
 {
+    public function test_create_from_values(): void
+    {
+        // Act
+        $angle = AngularDistance::createFromValues(
+            $this->randomDegrees()->value(),
+            $this->randomMinutes()->value(),
+            $this->randomSeconds()->value(),
+            $this->randomDirection()
+        );
+
+        // Assert
+        $this->assertInstanceOf(AngularDistance::class, $angle);
+    }
+
     #[TestDox("can be created from a decimal number.")]
     public function test_create_from_decimal(): void
     {
@@ -58,9 +96,6 @@ class AngularDistanceTest extends TestCase
         $angle = AngularDistance::createFromDecimal($decimal);
 
         // Assert
-        $this->assertEquals(
-            $decimal, 
-            $angle->toFloat(3)
-        );
+        $this->assertInstanceOf(AngularDistance::class, $angle);
     }
 }