+++ tests/Unit/Builders/AngularDistance/FromSexagesimalTest.php
@@ -1,0 +1,76 @@
+<?php
+namespace MarcoConsiglio\Goniometry\Tests\Unit\Builders\AngularDistance;
+
+use MarcoConsiglio\Goniometry\Angle;
+use MarcoConsiglio\Goniometry\Builders\Angle\FromSexadecimal;
+use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromSexadecimal as AngularDistanceFromSexadecimal;
+use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromSexagesimal;
+use MarcoConsiglio\Goniometry\Degrees;
+use MarcoConsiglio\Goniometry\Minutes;
+use MarcoConsiglio\Goniometry\Random\Generator\Angle as AngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\NegativeAngle as NegativeAngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\NegativeSexadecimal as NegativeSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\PositiveAngle as PositiveAngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\PositiveSexadecimal as PositiveSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\RelativeAngle as RelativeAngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Validator\FloatValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\NegativeSexadecimal as NegativeSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\PositiveSexadecimal as PositiveSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\RelativeSexadecimal as RelativeSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Seconds;
+use MarcoConsiglio\Goniometry\SexadecimalAngularDistance;
+use MarcoConsiglio\Goniometry\SexadecimalDegrees;
+use MarcoConsiglio\Goniometry\SexagesimalDegrees;
+use MarcoConsiglio\Goniometry\Tests\TestCase;
+use MarcoConsiglio\Goniometry\Traits\WithAngleFaker;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\UsesClass;
+use PHPUnit\Framework\Attributes\UsesTrait;
+
+#[CoversClass(FromSexagesimal::class)]
+#[UsesClass(Angle::class)]
+#[UsesClass(AngleGenerator::class)]
+#[UsesClass(AngularDistanceFromSexadecimal::class)]
+#[UsesClass(Degrees::class)]
+#[UsesClass(FloatValidator::class)]
+#[UsesClass(FromSexadecimal::class)]
+#[UsesClass(FromSexagesimal::class)]
+#[UsesClass(Minutes::class)]
+#[UsesClass(NegativeAngleGenerator::class)]
+#[UsesClass(NegativeSexadecimalGenerator::class)]
+#[UsesClass(PositiveAngleGenerator::class)]
+#[UsesClass(PositiveSexadecimalGenerator::class)]
+#[UsesClass(PositiveSexadecimalValidator::class)]
+#[UsesClass(NegativeSexadecimalValidator::class)]
+#[UsesClass(RelativeAngleGenerator::class)]
+#[UsesClass(RelativeSexadecimalValidator::class)]
+#[UsesClass(Seconds::class)]
+#[UsesClass(SexadecimalAngularDistance::class)]
+#[UsesClass(SexadecimalDegrees::class)]
+#[UsesClass(SexagesimalDegrees::class)]
+#[UsesTrait(WithAngleFaker::class)]
+class FromSexagesimalTest extends TestCase
+{
+    public function test_can_create_an_angular_distance_from_sexagesimal_degrees(): void
+    {
+        // Arrange
+        $angle = $this->randomAngle();
+        $sexadecimal = new SexadecimalAngularDistance($angle->toSexadecimalDegrees()->value);
+        $builder = new FromSexagesimal(
+            $angle->degrees->value(),
+            $angle->minutes->value(),
+            $angle->seconds->value(),
+            $angle->direction
+        );
+
+        // Act
+        $result = $builder->fetchData();
+        $sexadecimal_angular_distance = $result[1];
+
+        // Assert
+        $this->assertEquals(
+            $sexadecimal->value(precision: 3),
+            $sexadecimal_angular_distance->value(precision: 3)
+        );
+    }
+}